# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enjin Blockchain REST API",
    "description": "REST API for Enjin and Canary Relaychain and Matrixchain. Forked from Parity Technologies' Polkadot REST API.",
    "contact": {
      "url": "https://github.com/enjin/blockchain-rest-api"
    },
    "license": {
      "name": "GPL-3.0-or-later"
    },
    "version": "0.3.2"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Localhost"
    }
  ],
  "paths": {
...
    "/v1/accounts/{accountId}/balance-info": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Account balance info",
        "description": "Returns balance information for a given account including free, reserved, and locked balances.",
        "operationId": "get_balance_info",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "SS58-encoded account address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "at",
            "in": "query",
            "description": "Block hash or number to query at",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "useRcBlock",
            "in": "query",
            "description": "Treat 'at' as relay chain block identifier",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "token",
            "in": "query",
            "description": "Token symbol for chains with multiple tokens",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "denominated",
            "in": "query",
            "description": "When true, denominate balances using chain decimals",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account or block parameter"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
...
    "/v1/rc/accounts/{accountId}/balance-info": {
      "get": {
        "tags": [
          "rc"
        ],
        "summary": "RC get balance info",
        "description": "Returns balance information for a given account on the relay chain.",
        "operationId": "get_balance_info",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "SS58-encoded account address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "at",
            "in": "query",
            "description": "Block identifier (number or hash)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "description": "Token symbol (defaults to native token)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "denominated",
            "in": "query",
            "description": "Denominate balances using chain decimals",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcBalanceInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account address"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Relay chain not configured"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceInfoResponse": {
        "type": "object",
        "description": "Response for GET /accounts/{accountId}/balance-info",
        "required": [
          "at",
          "nonce",
          "tokenSymbol",
          "free",
          "reserved",
          "miscFrozen",
          "feeFrozen",
          "frozen",
          "transferable",
          "locks"
        ],
        "properties": {
          "ahTimestamp": {
            "type": [
              "string",
              "null"
            ]
          },
          "at": {
            "$ref": "#/components/schemas/BlockInfo"
          },
          "feeFrozen": {
            "type": "string",
            "description": "The amount that free may not drop below when withdrawing specifically for\ntransaction fee payment (legacy field, may be string message for newer runtimes)"
          },
          "free": {
            "type": "string",
            "description": "Free balance (not equivalent to spendable balance)"
          },
          "frozen": {
            "type": "string",
            "description": "Frozen balance (newer runtimes, may be string message for older runtimes)"
          },
          "locks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalanceLock"
            },
            "description": "Array of balance locks"
          },
          "miscFrozen": {
            "type": "string",
            "description": "The amount that free may not drop below when withdrawing for anything except\ntransaction fee payment (legacy field, may be string message for newer runtimes)"
          },
          "nonce": {
            "type": "string",
            "description": "Account nonce"
          },
          "rcBlockHash": {
            "type": [
              "string",
              "null"
            ]
          },
          "rcBlockNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "reserved": {
            "type": "string",
            "description": "Reserved balance"
          },
          "tokenSymbol": {
            "type": "string",
            "description": "Token symbol"
          },
          "transferable": {
            "type": "string",
            "description": "Calculated transferable balance using: free - max(maybeED, frozen - reserved)"
          }
        }
      },
      "BalanceLock": {
        "type": "object",
        "description": "Balance lock information",
        "required": [
          "id",
          "amount",
          "reasons"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "description": "Amount locked"
          },
          "id": {
            "type": "string",
            "description": "Lock identifier"
          },
          "reasons": {
            "type": "string",
            "description": "Lock reasons (Fee = 0, Misc = 1, All = 2)"
          }
        }
      },
      "BlockInfo": {
        "type": "object",
        "description": "Block information",
        "required": [
          "hash",
          "height"
        ],
        "properties": {
          "hash": {
            "type": "string"
          },
          "height": {
            "type": "string"
          }
        }
      },
...
      "RcBalanceInfoResponse": {
        "type": "object",
        "description": "Response for GET /rc/accounts/{accountId}/balance-info",
        "required": [
          "at",
          "nonce",
          "tokenSymbol",
          "free",
          "reserved",
          "miscFrozen",
          "feeFrozen",
          "frozen",
          "transferable",
          "locks"
        ],
        "properties": {
          "at": {
            "$ref": "#/components/schemas/BlockInfo"
          },
          "feeFrozen": {
            "type": "string",
            "description": "The amount that free may not drop below when withdrawing specifically for\ntransaction fee payment (legacy field, may be string message for newer runtimes)"
          },
          "free": {
            "type": "string",
            "description": "Free balance (not equivalent to spendable balance)"
          },
          "frozen": {
            "type": "string",
            "description": "Frozen balance (newer runtimes, may be string message for older runtimes)"
          },
          "locks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalanceLock"
            },
            "description": "Array of balance locks"
          },
          "miscFrozen": {
            "type": "string",
            "description": "The amount that free may not drop below when withdrawing for anything except\ntransaction fee payment (legacy field, may be string message for newer runtimes)"
          },
          "nonce": {
            "type": "string",
            "description": "Account nonce"
          },
          "reserved": {
            "type": "string",
            "description": "Reserved balance"
          },
          "tokenSymbol": {
            "type": "string",
            "description": "Token symbol"
          },
          "transferable": {
            "type": "string",
            "description": "Calculated transferable balance using: free - max(maybeED, frozen - reserved)"
          }
        }
      },
...
    }
  },
...
}